requirements.txt
discord.py
selenium
webdriver-manager
beautifulsoup4
openpyxl==3.1.2 
discord.py
pandas
python-dotenv>=1.0.0